[WEB-757] Fix linting errors
